requirements.txt
pandas
numpy
openai
python-dotenv
rapidfuzz
scikit-learn
gradio
Pillow
openpyxl